Sprint 1: Modular refactor, caching, versioning, README updates
