+++ gui/__init__.py
@@ -1,0 +1,4 @@
+# File: transcribe_audio_service/gui/__init__.py
+
+# This file marks the 'gui' directory as a Python package.
+# It can also be used to expose shared logic or simplify imports.
