+++ services/version.py
@@ -1,0 +1,2 @@
+# File: services/version.py
+__version__ = "1.1.0"
